+++ modules/tinymce/src/core/main/ts/api/dom/RangeUtils.ts
@@ -1,5 +1,6 @@
 import { Fun } from '@ephox/katamari';
 
+import * as ExpandRange from '../../fmt/ExpandRange';
 import * as CaretRangeFromPoint from '../../selection/CaretRangeFromPoint';
 import * as NormalizeRange from '../../selection/NormalizeRange';
 import * as RangeCompare from '../../selection/RangeCompare';
@@ -13,6 +14,7 @@ interface RangeUtils {
   walk: (rng: Range, callback: (nodes: Node[]) => void) => void;
   split: (rng: Range) => RangeLikeObject;
   normalize: (rng: Range) => boolean;
+  expand: (rng: Range, options?: { type: 'word' }) => Range;
 }
 
 /**
@@ -60,9 +62,31 @@ const RangeUtils = (dom: DOMUtils): RangeUtils => {
     );
   };
 
+  /**
+   * Returns a range expanded around the entire word the provided selection was collapsed within.
+   *
+   * @method expand
+   * @param {Range} rng The initial range to work from.
+   * @param {Object} options Optional options provided to the expansion. Defaults to { type: 'word' }
+   * @return {Range} Returns the expanded range.
+   */
+  const expand = (rng: Range, options: { type: 'word' } = { type: 'word' }): Range => {
+    if (options.type === 'word') {
+      const rangeLike = ExpandRange.expandRng(dom, rng, [{ inline: 'span' }]);
+      const newRange = dom.createRng();
+      newRange.setStart(rangeLike.startContainer, rangeLike.startOffset);
+      newRange.setEnd(rangeLike.endContainer, rangeLike.endOffset);
+
+      return newRange;
+    }
+
+    return rng;
+  };
+
   return {
     walk,
     split,
+    expand,
     normalize
   };
 };