+++ modules/tinymce/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- New `expand` function added to `tinymce.selection` which expands the selection around the nearest word. #TINY-9001
+- New `expand` function added to `tinymce.dom.RangeUtils` to return a new range expanded around the nearest word. #TINY-9001
+
 ## 6.2.0 - 2022-09-08
 
 ### Added